[UPDATE][REFACTOR] Added makefile and changed to header-based infrastructure
added ore_algebra.h, and separated GF coefficients into gf_coefficients.h
resp. gf_coefficients.c. Added makefile, so that the respective linking and compiling
works properly.
